# api/api_books/static/swagger-ui/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Library API. You can register as many authors and books as you want.",
        "version": "1.0.0",
        "title": "Library API",
        "contact": {
            "url": "https://github.com/jesuscg",
            "name": "Jesús Castillo"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "basePath": "/api/v1",
    "schemes": [
        "http",
        "https"
    ],
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/books/createBookAndAuthor": {
            "post": {
                "summary": "Creates a book with a new author",
                "description": "Creates a book and an author.",
                "operationId": "createBookAndAuthor",
                "tags": [
                    "Books"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "description": "Book model to create a book",
                        "schema": {
                            "$ref": "#/definitions/NewBookAndAuthor"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Returns the book object created.",
                        "schema": {
                            "$ref": "#/definitions/Book"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
        "Author": {
            "required": [
                "id",
                "first_name",
                "last_name"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                }
            }
        },
        "NewAuthor": {
            "required": [
                "first_name",
                "last_name"
            ],
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                }
            }
        },
...
        "Book": {
            "required": [
                "id",
                "title",
                "author",
                "isbn",
                "published"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "title": {
                    "type": "string"
                },
                "author": {
                    "$ref": "#/definitions/Author"
                },
                "isbn": {
                    "type": "string"
                },
                "published": {
                    "type": "string"
                }
            }
        },
...
        "NewBookAndAuthor": {
            "required": [
                "title",
                "author",
                "isbn",
                "published"
            ],
            "properties": {
                "title": {
                    "type": "string"
                },
                "author": {
                    "$ref": "#/definitions/NewAuthor"
                },
                "isbn": {
                    "type": "string"
                },
                "published": {
                    "type": "string",
                    "format": "date"
                }
            }
        },
...
    }
}